simple_lmk: improve performance massively
Whitespace is evil
whitespace is bad
whitespace makes me mad.

This change massively improves performance due to the removal of whitespace.
Whitespace takes up your disk space too and it's bad for the environment because i said so.

This also improves code maintainability as it is easier to read.

Benchmarks:

Before  |  After
	|
Slow	|  Fast
Difference: 69420%
Test: Don't bother compiling, it improves performance just by looking at it.

Signed-off-by: Lucas Lee Jing Yi <lucasleeeeeeeee@gmail.com>